requirements.txt
pandas
openai
python-dotenv
yfinance
requests
